[LCR-139] feat: 설문지 분석 기능 구현

+++ src/main/java/com/lgcns/domain/survey/dto/response/ChoiceStatsResponse.java
@@ -1,7 +1,7 @@
 package com.lgcns.domain.survey.dto.response;
 
-public record ChoiceStatsResponse(String choiceContent, int selectedCount, double ratio) {
-    public static ChoiceStatsResponse of(String choiceContent, int selectedCount, double ratio) {
+public record ChoiceStatsResponse(String choiceContent, Long selectedCount, Double ratio) {
+    public static ChoiceStatsResponse of(String choiceContent, Long selectedCount, Double ratio) {
         return new ChoiceStatsResponse(choiceContent, selectedCount, ratio);
     }
 }
